Fix: Team Canal not deleted at end game
Bug: Team Canal are not deleted when host leave, but player are removed from channel when leaving

Cause: When host leaves, the async nature of the deleteRoomCanal make it so when trying to get teams inside the the method, the room has already been deleted in the Socket.HostLeft Handler. Therefore there is no teams (teams == undefined) and teams Canal are not deleted

Fix: Get teams Ids directly in socket event handler to be ensure to get teams Id before deleting room

Tested with 4 team games and 1 classic game => works!