Fix: Same score Win/Loss distribution
Bug: If all teams get the same score or all player get the same score, they both are considered as winner and loser. Therefore they receive winner and loser compensation and win/loss in game history.

Cause: In calculateExtremesScore for teams and player, we dont consider the fact that all might have the score.

Fix: Verify the presence of a player of a team id in both loser and winner bracket and remove them from loser bracket if the totalscore of all team or player is more than 0 (which means that some points where attributed).

Test 5 classic qrl only game with same score, 5 teams game qrl only with 0 for everyone and 2 game as usual for team and normal mode